% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_sd_fxn.R
\name{avg_sd_fxn}
\alias{avg_sd_fxn}
\title{Get formatted character vector of the average +/- SD.}
\usage{
avg_sd_fxn(x, n.digits = 1, include.n = FALSE)
}
\arguments{
\item{x}{Numeric Vector}

\item{n.digits}{number of digits after the decimal point. Default is 1.}

\item{include.n}{include sample size? Default is false.}
}
\description{
Get formatted character vector of the average +/- SD.
}
