% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ft.R
\name{clean_ft}
\alias{clean_ft}
\title{Clean Raw Feature Table}
\usage{
clean_ft(raw_ft, samples)
}
\arguments{
\item{raw_ft}{Raw LC-MS Feature Table}

\item{samples}{Which to select? One of: "all", "sol", "q3". In future: "duplicates"}
}
\description{
Clean Raw Feature Table
}
