% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effest_ci.R
\name{effest_ci}
\alias{effest_ci}
\title{Get formatted character with effect estimate and 95\% CI}
\usage{
effest_ci(eff_est, lcl, ucl, n.digits = 2)
}
\arguments{
\item{eff_est}{Numeric Vector}

\item{lcl}{Numeric Vector}

\item{ucl}{Numeric Vector}

\item{n.digits}{number of digits after the decimal point. Default is 2.}
}
\description{
Get formatted character with effect estimate and 95\% CI
}
