% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fungm.R
\name{fungm}
\alias{fungm}
\title{Get formatted character with geometric mean and 95\% CI}
\usage{
fungm(x, n.digits = 3, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric Vector}

\item{n.digits}{number of digits after the decimal point. Default is 3.}

\item{na.rm}{Remove NAs from calculation? Default is FALSE.}
}
\description{
Get formatted character with geometric mean and 95\% CI
}
