% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fungsd.R
\name{fungsd}
\alias{fungsd}
\title{Get formatted character with geometric mean and sd}
\usage{
fungsd(x, n.digits = 1, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric Vector}

\item{n.digits}{number of digits after the decimal point. Default is 1.}

\item{na.rm}{Remove NAs from calculation? Default is FALSE.}
}
\description{
Get formatted character with geometric mean and sd
}
