% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_iqr_fxn.R
\name{median_iqr_fxn}
\alias{median_iqr_fxn}
\title{Get formatted character vector of the Mean (IQR).}
\usage{
median_iqr_fxn(x, n.digits = 1)
}
\arguments{
\item{x}{Numeric Vector}

\item{n.digits}{number of digits after the decimal point. Default is 1.}
}
\description{
Get formatted character vector of the Mean (IQR).
}
