% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwas.R
\name{mwas_all_outcomes}
\alias{mwas_all_outcomes}
\title{MWAS Function (WORK IN PROGRESS)}
\usage{
mwas_all_outcomes(metab_dat, exp_cov_dat, exposure_names)
}
\arguments{
\item{metab_dat}{Data frame, where the first column is the unique id, and the rest of the columns are metabolites}

\item{exp_cov_dat}{data frame which contains one column with the unique id as well as all of the covariates}

\item{covariates}{vector containing the covariates for}
}
\description{
MWAS Function (WORK IN PROGRESS)
}
