% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npct.R
\name{npct}
\alias{npct}
\title{Get formatted character with n and percent from a factor, character, or numeric}
\usage{
npct(x, level_of_interest, n.digits = 1)
}
\arguments{
\item{x}{Values}

\item{level_of_interest}{Value of interest, for example, "Male"}

\item{n.digits}{number of digits after the decimal point. Default is 1.}
}
\description{
Get formatted character with n and percent from a factor, character, or numeric
}
