% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owas.R
\name{owas}
\alias{owas}
\title{FAST OWAS Function}
\usage{
owas(df, var, omics, covars, var_exposure_or_outcome, model_type = "linear")
}
\arguments{
\item{df}{Dataset}

\item{var}{Name of the variable of interest- this is usually either an exposure 
variable or an outcome variable}

\item{omics}{Names of all omics features in the dataset}

\item{covars}{Names of covariates (can be NULL)}

\item{var_exposure_or_outcome}{Is the variable of interest an exposure 
(independent variable) or outcome (dependent variable)? Must be either
"exposure" or "outcome"}

\item{model_type}{"linear" for linear models (via lm) or "logistic" for 
logistic (via glm)}
}
\description{
FAST OWAS Function
}
