% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_act.R
\name{p_act}
\alias{p_act}
\title{Generalized P-Act Function}
\usage{
p_act(p_values, x_vars, y_vars = NA, covariates, alpha = 0.05)
}
\arguments{
\item{p_values}{Data frame with 3 columns. File should contain one row for 
each test. Missingness is not allowed. 
 
  Column 1: dpdt_var names (outcomes)   
   
  Column 2: ind_var names   
  
  Column 3: The p-value from each test}

\item{covariates}{Data frame with one column for each covariate, 
and one row for each individual. Column names should be the names of the 
covariates. Not necessary if no covariates used; Not necessary if only 
one independent variable is considered. Missing values are ok.
Must have same # of individuals as @ind_vars.}

\item{alpha}{Used to determine when to stop sequential testing. 
If not specified, default value of .05 is used.}

\item{dpdt_vars}{traits.txt: 
  Not necessary if only one trait is considered.
One column for each trait. 
One row for each individual plus header row of trait names.  
If covariates are used, need trait values residualized on covariates.
Missing values coded as NA.  
Must have same # of individuals as metabolite.txt.}

\item{ind_vars}{metabolite.txt/outcomes.txt: 
  One column for each metabolite score. 
One row for each individual plus a header row of metabolite labels.  
Missing values must be coded as NA.
Not necessary if only one marker is considered.}
}
\value{
a data frame containing 4 columns: y_vars, x_vars, original p values, 
and p-act corrected p values.
}
\description{
P_act was designed for correction of GWAS analysis. Here, 
the code has been modified for the following generalized scenario:          
dpdt_var ~ ind_var + covariates      
Where:          
     dpdt_vars is a set of correlated outcomes/traits;       
     ind_vars is a set of potentially correlated exposures/metabolites/genes/etc..;        
     and covariates is a set of covariates for the analysis.           
This function can not handle more than 1,000 corrections.
}
