% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qntle_fxn.R
\name{qntle_fxn}
\alias{qntle_fxn}
\title{Get quantiles for exposure concentrations. Returns a charachter vector.}
\usage{
qntle_fxn(x, pct)
}
\arguments{
\item{x}{Vector of exposure concentrations}

\item{pct}{percentile cutoff}
}
\description{
Get quantiles for exposure concentrations. Returns a charachter vector.
}
