% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_names_cv.R
\name{reformat_names}
\alias{reformat_names}
\title{Reformat author names for CV}
\usage{
reformat_names(name_string)
}
\arguments{
\item{name_string}{String of author names, in the format "first last#, ..."}
}
\description{
Reformat author names for CV
}
