% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_exposures.R
\name{rename_exposures}
\alias{rename_exposures}
\title{Rename exposures, including PFAS and lipophilic pops.}
\usage{
rename_exposures(
  exposure_names,
  include_asterisk = FALSE,
  arrange_by_class = FALSE
)
}
\arguments{
\item{exposure_names}{character vector of lowercase PFAS or lipophilic POPs names}

\item{include_asterisk}{Include asterisk for exposures which will likely be dichotomous? May need to be altered for different projects based on percent below lod.}

\item{arrange_by_class}{Order factor levels by chemical class?}
}
\description{
Rename exposures, including PFAS and lipophilic pops.
}
