% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superpathway_key.R
\docType{data}
\name{superpathway_key}
\alias{superpathway_key}
\title{Pathway to superpathway key}
\format{
## `superpathway_key`
A data frame with 99 rows and 3 columns
}
\usage{
superpathway_key
}
\description{
Pathway to superpathway key
}
\keyword{datasets}
